data type for retention duration and sample description changes

+++ storage/api/Storage.Samples.Tests/BucketSetSoftDeletePolicyTest.cs
@@ -30,7 +30,7 @@ public class BucketSetSoftDeletePolicyTest
         BucketSetSoftDeletePolicySample bucketSetSoftDeletePolicy = new BucketSetSoftDeletePolicySample();
         var bucketName = _fixture.GenerateBucketName();
         _fixture.CreateBucket(bucketName, multiVersion: false, softDelete: true, registerForDeletion: true);
-        long retentionDuration = 10;
+        int retentionDuration = 10;
         // Set soft-delete policy for the bucket.
         var bucket = bucketSetSoftDeletePolicy.BucketSetSoftDeletePolicy(bucketName, retentionDuration);
         long retentionDurationInSeconds = retentionDuration * (24 * 60 * 60);
